chore(29/12/2023): refactor e2e test cases
