+++ test/devices/devices.spec.ts
@@ -1,73 +1,21 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import {
-  BadRequestException,
-  HttpStatus,
-  INestApplication,
-  ValidationPipe,
-} from "@nestjs/common";
-import cookieParser from "cookie-parser";
+import { HttpStatus, INestApplication } from "@nestjs/common";
 import request from "supertest";
 import {
   AuthLoginInputModal,
   AuthRegistrationInputModal,
 } from "../../src/features/roles/public/auth/api/auth.models";
-import { AppModule } from "../../src/app.module";
-import { HttpExceptionFilter } from "../../src/exception.filter";
-import { useContainer } from "class-validator";
 import { DeleteDataController } from "../../src/features/infrstructura/deleting-all-data";
-import { Request, Response } from "express";
 import { v4 as uuidv4 } from "uuid";
+import { initTestApp } from "../init.app";
+import { mockRequest, mockResponse } from "../__test-data__";
 
-const mockRequest = {
-  headers: {
-    "user-agent": "device name",
-  },
-} as unknown as Request;
-
-const mockResponse = {
-  cookie: jest.fn(),
-  status: jest.fn(() => mockResponse),
-  send: jest.fn(() => true),
-} as unknown as Response;
-
-describe("Devices E2E", () => {
+describe("Devices", () => {
   let app: INestApplication;
   let deleteDataController: DeleteDataController;
 
   beforeAll(async () => {
-    const moduleFixture: TestingModule = await Test.createTestingModule({
-      imports: [AppModule],
-    }).compile();
-
-    app = moduleFixture.createNestApplication();
-    useContainer(app.select(AppModule), { fallbackOnErrors: true });
-    app.use(cookieParser());
-
-    app.useGlobalPipes(
-      new ValidationPipe({
-        stopAtFirstError: true,
-        transform: true,
-        transformOptions: { enableImplicitConversion: true },
-        exceptionFactory: (errors) => {
-          const errorsForProperty: any[] = [];
-
-          errors.forEach((e) => {
-            const constrainKey = Object.keys(e.constraints!);
-            constrainKey.forEach((cKey) => {
-              errorsForProperty.push({
-                field: e.property,
-                message: e.constraints![cKey],
-              });
-            });
-          });
-
-          throw new BadRequestException(errorsForProperty);
-        },
-      })
-    );
-    app.useGlobalFilters(new HttpExceptionFilter());
+    app = await initTestApp();
     await app.init();
-
     deleteDataController = app.get<DeleteDataController>(DeleteDataController);
   });
